% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{bgbb_d_regularization}
\alias{bgbb_d_regularization}
\title{Derivative of regularization function for BG/BB model}
\usage{
bgbb_d_regularization(params, lambda)
}
\arguments{
\item{params}{A named list of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{lambda}{The regularization factor}
}
\description{
Derivative of regularization function for BG/BB model
}
