% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{bgbb_loglikelihood}
\alias{bgbb_loglikelihood}
\title{Log-Likelihood function for BG/BB model}
\usage{
bgbb_loglikelihood(data, params)
}
\arguments{
\item{data}{An integer matrix where each row consists of the
recency/frequency statistics for a single user, in the order (x, n, m)}

\item{params}{A named list of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}
}
\description{
Log-Likelihood function for BG/BB model
}
