% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{bgbb_regularization}
\alias{bgbb_regularization}
\title{Regularization function for BG/BB model}
\usage{
bgbb_regularization(params, lambda)
}
\arguments{
\item{params}{A named list of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{lambda}{The regularization factor}
}
\description{
Regularization function for BG/BB model
}
