% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.r
\name{calculate_sss}
\alias{calculate_sss}
\title{Calculate sensitivity + specificity statistic for a given result set, at a
given prediction probability threshold.}
\usage{
calculate_sss(threshold = 0.5, comparison_data)
}
\arguments{
\item{threshold}{A number between 0 and 1. Any user whose estimated return
probability exceeds this threshold is predicted to return next week.}

\item{comparison_data}{A data.frame of the form (user_id,
return_probability, showed_up). return_probability is the users' estimated
return probability, and showed_up is an indicator of whether they actually
showed up (1 = they showed up)}
}
\value{
A number between -1  and 1 representing sensitivity + specificity
- 1. Higher numbers represent accurate forecasts
}
\description{
Calculate sensitivity + specificity statistic for a given result set, at a
given prediction probability threshold.
}
