% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.r
\name{calculateConfusionMatrix}
\alias{calculateConfusionMatrix}
\title{Calculate confusion matrix for a given set of comparison data}
\usage{
calculateConfusionMatrix(threshold = 0.5, comparisonData)
}
\arguments{
\item{threshold}{A number between 0 and 1. Any user whose estimated return
probability exceeds this threshold is predicted to return next week.}

\item{comparisonData}{A data.frame of the form (user_id,
return_probability, showed_up). return_probability is the users' estimated
return probability, and showed_up is an indicator of whether they actually
showed up (1 = they showed up)}
}
\value{
A data frame of the form (prediction, showed_up, user_count) that
gives the number of users that fall into each category
}
\description{
Calculate confusion matrix for a given set of comparison data
}
