% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.r
\name{calculateRecencyFrequency}
\alias{calculateRecencyFrequency}
\title{Calculate recency and frequency statistics from session duration data}
\usage{
calculateRecencyFrequency(sessDurData)
}
\arguments{
\item{sessDurData}{A data.frame of the form (user_id,
active_week_start_date) that gives the weeks in which each user was active.}
}
\value{
A data.frame of the form (user_id, x, n, m). Here, x represents the
number of repeat transactions made by the user, n represents the number of
repeat transaction opportunities that the user had, and m represents the
index of the most recent transaction. (M is also called t_x in the
literature.)
}
\description{
Calculate recency and frequency statistics from session duration data
}
