% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.r
\name{calculateROC}
\alias{calculateROC}
\title{Calculate ROC curve for a given set of comparison data}
\usage{
calculateROC(threshold = 0.5, comparisonData)
}
\arguments{
\item{threshold}{A number between 0 and 1. Any user whose estimated return
probability exceeds this threshold is predicted to return next week.}

\item{comparisonData}{A data.frame of the form (user_id,
return_probability, showed_up). return_probability is the users' estimated
return probability, and showed_up is an indicator of whether they actually
showed up (1 = they showed up)}
}
\value{
A length-2 numeric vector of the form (1 - specificity, sensitivity)
giving the coordinates in ROC space corresponding to the given prediction
threshold.
}
\description{
Calculate ROC curve for a given set of comparison data
}
