% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{comparison_test_data_homogeneous}
\alias{comparison_test_data_homogeneous}
\title{Data frame for testing the output of the calculate_sps function}
\format{data.frame}
\usage{
comparison_test_data_homogeneous
}
\description{
Data frame for testing the output of the calculate_sps function
}
\keyword{datasets}
