% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{cost_d}
\alias{cost_d}
\title{Derivative of cost function}
\usage{
cost_d(params, data, lambda = 0)
}
\arguments{
\item{params}{A numeric vector of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{data}{An integer matrix where each row consists of the
recency/frequency statistics for a single user, in the order (x, n, m)}

\item{lambda}{The regularization factor}
}
\description{
Derivative of cost function
}
