% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{estimateExpectedTransactions}
\alias{estimateExpectedTransactions}
\title{Estimate the expected number of active weeks that a user will have 
out of the next n weeks, given their recency/frequency statistics 
and model parameters alpha through gamma}
\usage{
estimateExpectedTransactions(n_star, params, data)
}
\arguments{
\item{n_star}{The number of}

\item{params}{A numeric vector of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{data}{An integer vector consisting of the recency/frequency statistics
for a single user, in the order (x, n, m)}
}
\description{
Estimate the expected number of active weeks that a user will have 
out of the next n weeks, given their recency/frequency statistics 
and model parameters alpha through gamma
}
