% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{estimateReturnProbability}
\alias{estimateReturnProbability}
\title{Estimate the probability that a user will show up this week,
given their recency/frequency statistics and model parameters alpha
through gamma}
\usage{
estimateReturnProbability(params, data)
}
\arguments{
\item{params}{A numeric vector of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{data}{An integer vector consisting of the recency/frequency statistics
for a single user, in the order (x, n, m)}
}
\description{
Estimate the probability that a user will show up this week,
given their recency/frequency statistics and model parameters alpha
through gamma
}
