% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.r
\name{fetchSessDurData}
\alias{fetchSessDurData}
\title{Fetch session duration data for a given group of users}
\usage{
fetchSessDurData(userGroup = NULL, runDate = as.numeric(gsub(pattern = "-",
  replacement = "", x = Sys.Date())), con = redshift_connection$con)
}
\arguments{
\item{userGroup}{A numeric vector of user_ids. Only user_ids belonging to this
set will be included in the analysis.}

\item{runDate}{A numeric date id in the form yyyymmdd. All data after this
date will be dropped.}

\item{con}{A database connection object to fetch the results from.}
}
\value{
data.frame of the form (user_id, active_week_start_date)
}
\description{
Fetch session duration data for a given group of users
}
