% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.r
\name{getFirstChamp}
\alias{getFirstChamp}
\title{Categorize users according to their first champion (REVEAL, FamilyLife, Other)}
\usage{
getFirstChamp(userGroup = NULL, runDate = as.numeric(gsub(pattern = "-",
  replacement = "", x = Sys.Date())), csvLoc = NULL,
  con = redshift_connection$con)
}
\arguments{
\item{userGroup}{A numeric vector of user_ids. Only user_ids belonging to this
set will be included in the analysis.}

\item{runDate}{A numeric date id in the form yyyymmdd. All data after this
date will be dropped.}

\item{csvLoc}{A string containing the location of a csv file containing
session duration data. If left NULL, data will be fetched from Redshift.}

\item{con}{A database connection object to fetch the results from.}
}
\value{
data.frame of the form (user_id, first_champ)
}
\description{
Categorize users according to their first champion (REVEAL, FamilyLife, Other)
}
