% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_get_user_ids}
\alias{query_get_user_ids}
\title{A string containing the user ids query.}
\format{A length-one character vector.}
\usage{
query_get_user_ids
}
\description{
A string containing the user ids query.
}
\keyword{datasets}
