% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_session_duration_data}
\alias{query_session_duration_data}
\title{A string containing the session duration data query.}
\format{A length-one character vector.}
\usage{
query_session_duration_data
}
\description{
A string containing the session duration data query.
}
\keyword{datasets}
