% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_session_duration_data_sub}
\alias{query_session_duration_data_sub}
\title{A string containing the session duration data query, with placeholders to
substitue userGroup and runDate queries.}
\format{A length-one character vector.}
\usage{
query_session_duration_data_sub
}
\description{
A string containing the session duration data query, with placeholders to
substitue userGroup and runDate queries.
}
\keyword{datasets}
