% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_user_first_champ}
\alias{query_user_first_champ}
\title{A string containing the user_first_champ query.}
\format{A length-one character vector.}
\usage{
query_user_first_champ
}
\description{
A string containing the user_first_champ query.
}
\keyword{datasets}
