% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_user_first_champ_sub}
\alias{query_user_first_champ_sub}
\title{A string containing the user_first_champ query, with placeholders to
substitue userGroup and runDate queries.}
\format{A length-one character vector.}
\usage{
query_user_first_champ_sub
}
\description{
A string containing the user_first_champ query, with placeholders to
substitue userGroup and runDate queries.
}
\keyword{datasets}
