% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{recency_frequency_mini_test}
\alias{recency_frequency_mini_test}
\title{A data.frame containing the recency/frequency statistics for
user_id_sample_mini}
\format{numeric}
\usage{
recency_frequency_mini_test
}
\description{
A data.frame containing the recency/frequency statistics for
user_id_sample_mini
}
\keyword{datasets}
