% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{regularization}
\alias{regularization}
\title{Regularization function for BG/BB model}
\usage{
regularization(params, lambda = 0)
}
\arguments{
\item{params}{A vector of parameters for the log likelihood
function. c(alpha, beta, gamma, delta)}

\item{lambda}{The regularization factor}
}
\description{
Regularization function for BG/BB model
}
