% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{regularization_d}
\alias{regularization_d}
\title{Derivative of regularization function for BG/BB model}
\usage{
regularization_d(params, lambda = 0)
}
\arguments{
\item{params}{A numeric vector of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}

\item{lambda}{The regularization factor}
}
\description{
Derivative of regularization function for BG/BB model
}
