% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{test_data_cRF_NA}
\alias{test_data_cRF_NA}
\title{A data frame containing the session duration data that produces the NA error
caused by calculateRecencyFrequencyData, with rundate = "2016-10-05".}
\format{data.frame}
\usage{
test_data_cRF_NA
}
\description{
A data frame containing the session duration data that produces the NA error
caused by calculateRecencyFrequencyData, with rundate = "2016-10-05".
}
\keyword{datasets}
