% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{trainModel}
\alias{trainModel}
\title{Function to train model}
\usage{
trainModel(data, paramsInit = c(1, 1, 1, 1), ...)
}
\arguments{
\item{data}{An integer matrix where each row consists of the
recency/frequency statistics for a single user, in the order (x, n, m)}

\item{paramsInit}{A length-4 numeric vector of initial parameters for 
the log likelihood function. (alpha, beta, gamma, delta)}

\item{...}{Additional arguments to pass to stats::optim}
}
\description{
Function to train model
}
