% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{user_id_sample_mini}
\alias{user_id_sample_mini}
\title{A random sample of 10 user ids.}
\format{numeric}
\usage{
user_id_sample_mini
}
\description{
A random sample of 10 user ids.
}
\keyword{datasets}
