% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{loglikelihood_d0}
\alias{loglikelihood_d0}
\title{Derivative of log-Likelihood function for BG/BB model (single user)}
\usage{
loglikelihood_d0(data, params)
}
\arguments{
\item{data}{An integer vector consisting of the recency/frequency statistics
for a single user, in the order (x, n, m)}

\item{params}{A numeric vector of parameters for the log likelihood
function. (alpha, beta, gamma, delta)}
}
\description{
Derivative of log-Likelihood function for BG/BB model (single user)
}
